refactor: flatten net::peers module into net (#4691)
