+++ fedimint-server/src/lib.rs
@@ -38,7 +38,7 @@ use fedimint_core::config::P2PMessage;
 use fedimint_core::db::{Database, DatabaseTransaction, IDatabaseTransactionOpsCoreTyped as _};
 use fedimint_core::epoch::ConsensusItem;
 use fedimint_core::module::ApiAuth;
-use fedimint_core::net::peers::DynP2PConnections;
+use fedimint_core::net::DynP2PConnections;
 use fedimint_core::task::{TaskGroup, sleep};
 use fedimint_core::util::SafeUrl;
 use fedimint_logging::LOG_CONSENSUS;
@@ -59,7 +59,7 @@ use crate::config::ConfigGenSettings;
 use crate::config::io::write_server_config;
 use crate::config::setup::{ConfigGenOutcome, SetupApi};
 use crate::db::{ServerInfo, ServerInfoKey};
-use crate::fedimint_core::net::peers::IP2PConnections;
+use crate::fedimint_core::net::IP2PConnections;
 use crate::metrics::initialize_gauge_metrics;
 use crate::net::api::announcement::start_api_announcement_service;
 use crate::net::api::pkarr_publish::start_pkarr_publish_service;
